Remove constraint on individual.number+individual.herd_id
This constraint makes no sense and does not add anything.  If the
constraint was to keep individual.number unique within a genebank,
then it does not do that.  There is no individual.genebank_id field so
numbers can't be made unique within a genebank in this way (usin herd_id
instead of genebank_id misses the point).