Remove the unique and non-NULL constraints on individual.certificate again
This is a bit controversial as we recently introduce the constraints on
this field.  However, for the same reasons herd.number can't be made
unique (there may be duplicates when looking across genebanks, and we
don't have an individual.genebank_id connection), individual.certificate
can't be made unique.  Also, there is no way around missing
certificates.  Some individuals simply have no certificates, and often
have this fact annetated as part of their data.

So, allow individual.certificate to be non-unique and NULL.

We may well want to have some sort of health check for the herd owners
later, so that they may find instances of empty or seemingly duplicated
certificates in the data.  A similar health check could be implemented
for the numbers.