Copy configuration (if any) to temporary storage and use that
Currently, the WSGI code will try to provide a configuration if
none is provided. If we point the configuration environment variable
(HERDBOOK_CONFIG) to the /config directory, the server may not be able
to create it there for permissions reasons (or r/o).

Copy any configuration to /tmp and use that instead.