Switch over to use new environment variable file
Use `../.docker/database-variables.env` in place of `../app/config.ini`.

Since this is a shell script file which defines a number of shell
variables, we source it and then transfer the values that we'd want to
use into the environment variables that `psql` will look at.

We require that `POSTGRES_DB`, `POSTGRES_USER`, `POSTGRES_PORT`,
`POSTGRES_HOST`, and `POSTGRES_PASSWORD` are defined in the file, but we
allow all but the first two of these to be empty.