Add herd start date
Read the herd "start date" from the auxilliary data file and insert it
into the `herd` table.  This requires us to first remove any data that
is not on the `YYYY-MM-DD` form, and then change the column format on
the `g_data2` table for the `"Start"` column, before simlpy doing an
`UPDATE` as for the other pieces of auxilliary data.